backend/requirements.txt
fastapi
uvicorn
sqlalchemy
databases
pydantic
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
email-validator
